pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-agent-framework"
version = "0.1.0"
description = "Deterministic loop-driven Codex CLI orchestration for project delivery."
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
keywords = ["codex", "cli", "agent", "automation", "orchestration"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Quality Assurance",
]
urls = { Homepage = "https://github.com/coolrazor007/project-agent-framework", Repository = "https://github.com/coolrazor007/project-agent-framework", Issues = "https://github.com/coolrazor007/project-agent-framework/issues" }
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.0",
  "twine>=5.1.1",
]

[project.scripts]
project-agent = "project_agent_framework.cli:main"
project-agent-framework = "project_agent_framework.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
project_agent_framework = [
  "resources/agents/*.md",
  "resources/demo_project/README.md",
  "resources/demo_project/PRD.md",
  "resources/demo_project/TODO.yml",
  "resources/demo_project/pyproject.toml",
  "resources/demo_project/data/*.yml",
  "resources/demo_project/src/inventory_report/*.py",
  "resources/demo_project/tests/.gitkeep",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
